ai_knowledge_assistant/requirements.txt
streamlit==1.31.0
sentence-transformers==2.3.1
PyPDF2==3.0.1
numpy==1.24.3
google-generativeai==0.3.2
torch==2.1.2
python-dotenv==1.0.0
